feat: Update Employee and Holiday resources with new fields and relationships - Enhanced Employee model with new fields: birth_date, payroll_type, employment_type, daily_rate, and face_descriptor. - Added leaves relationship to Employee model. - Updated EmployeeFactory to generate new fields and relationships. - Created EmployeeLeave resource and its management pages. - Added migration for employee_leaves table with necessary fields and relationships. - Updated existing migrations for employees, departments, schedules, and attendance days to reflect new requirements. - Created seeders for Branch, Deduction, Department, EmployeeDeduction, EmployeePerception, Holiday, and Schedule with relevant data.

+++ database/migrations/2025_05_20_132107_create_employee_deductions_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreignId('deduction_id')->constrained()->cascadeOnDelete(); // Relación con la tabla deductions
             $table->date('start_date')->nullable(); // Fecha de inicio de la deducción
             $table->date('end_date')->nullable(); // Fecha de fin de la deducción (opcional)
-            $table->integer('custom_amount')->nullable(); // Monto personalizado de la deducción (opcional)
+            $table->decimal('custom_amount', 10, 2)->nullable(); // Monto personalizado de la deducción (opcional)
             $table->unique(['employee_id', 'deduction_id', 'start_date'], 'emp_ded_sta_unique');
             $table->timestamps();
         });
